nbft: Turn nbft_info structs into fixed-sized arrays
This removes the ccan includes from the public headers.

No need for a pointer since the other elements in the struct are equally
valid at the same time and according to the specs the mac_addr value should
be set to a non-zero value.

Fix get_heap_obj() printf element size

strnlen() returns arch-dependent size_t

Sign-off-by: Tomas Bzatek <tbzatek@redhat.com>